add test for intros
